refactor(channels): move filterChannels to getChannelsWithReadAbility
